Break ground on @directus/stores (#18829)
* Setup @directus/stores

* Move appStore to @directus/stores

* Add changeset

* Initial version

---------
