+++ .changeset/clean-wasps-build.md
@@ -1,0 +1,6 @@
+---
+'@directus/stores': minor
+'@directus/app': patch
+---
+
+Move useAppStore to @directus/stores
