+++ app/src/layouts/map/options.vue
@@ -38,9 +38,9 @@
 </template>
 
 <script setup lang="ts">
-import { useAppStore } from '@/stores/app';
 import { getBasemapSources } from '@/utils/geometry/basemap';
 import { useSync } from '@directus/composables';
+import { useAppStore } from '@directus/stores';
 import { GeometryOptions, Item } from '@directus/types';
 import { toRefs } from 'vue';
 import { useI18n } from 'vue-i18n';
