+++ app/src/routes/login/login.test.ts
@@ -1,21 +1,20 @@
+import vIcon from '@/components/v-icon/v-icon.vue';
+import vImage from '@/components/v-image.vue';
+import vSelect from '@/components/v-select/v-select.vue';
+import MarkdownDirective from '@/directives/markdown';
+import { useServerStore } from '@/stores/server';
+import publicView from '@/views/public/public-view.vue';
+import { useAppStore } from '@directus/stores';
 import { createTestingPinia } from '@pinia/testing';
 import { mount } from '@vue/test-utils';
 import { GlobalMountOptions } from '@vue/test-utils/dist/types';
 import { setActivePinia } from 'pinia';
 import { beforeEach, expect, test, vi } from 'vitest';
 import { createI18n } from 'vue-i18n';
-
-import vIcon from '@/components/v-icon/v-icon.vue';
-import vImage from '@/components/v-image.vue';
-import vSelect from '@/components/v-select/v-select.vue';
-import MarkdownDirective from '@/directives/markdown';
-import { useAppStore } from '@/stores/app';
-import publicView from '@/views/public/public-view.vue';
 import ContinueAs from './components/continue-as.vue';
-import SsoLinks from './components/sso-links.vue';
 import LdapForm from './components/login-form/ldap-form.vue';
 import LoginForm from './components/login-form/login-form.vue';
-import { useServerStore } from '@/stores/server';
+import SsoLinks from './components/sso-links.vue';
 
 import LoginComponent from './login.vue';
