+++ app/src/routes/tfa-setup.vue
@@ -43,8 +43,8 @@
 <script setup lang="ts">
 import { useTFASetup } from '@/composables/use-tfa-setup';
 import { router } from '@/router';
-import { useAppStore } from '@/stores/app';
 import { useUserStore } from '@/stores/user';
+import { useAppStore } from '@directus/stores';
 import { User } from '@directus/types';
 import { nextTick, onMounted, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
