+++ app/src/views/private/components/sidebar-button.vue
@@ -15,8 +15,8 @@
 </template>
 
 <script setup lang="ts">
+import { useAppStore } from '@directus/stores';
 import { toRefs } from 'vue';
-import { useAppStore } from '@/stores/app';
 
 withDefaults(
 	defineProps<{
