+++ app/src/views/private/private-view.vue
@@ -100,7 +100,7 @@ import VResizeable, { ResizeableOptions } from '@/components/v-resizeable.vue';
 import { useLocalStorage } from '@/composables/use-local-storage';
 import { useTitle } from '@/composables/use-title';
 import { useWindowSize } from '@/composables/use-window-size';
-import { useAppStore } from '@/stores/app';
+import { useAppStore } from '@directus/stores';
 import { useUserStore } from '@/stores/user';
 import { useElementSize, useSync } from '@directus/composables';
 import { useEventListener } from '@vueuse/core';
