+++ packages/stores/src/index.ts
@@ -1,0 +1,1 @@
+export { useAppStore } from './app.js';
