+++ packages/stores/tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"extends": "@directus/tsconfig/node18-esm.json",
+	"compilerOptions": {
+		"lib": ["es2022", "DOM"]
+	}
+}
